requirements.txt
sentence-transformers==2.2.2
torch==2.1.1
transformers==4.35.0
faiss-cpu==1.7.4
pandas==2.1.3
numpy==1.26.2
tqdm==4.66.1
pyyaml==6.0.1
